requirements.txt
streamlit-audiorec
pyjwt
srt
pydub
python-dotenv
streamlit